Feature/build 64bit polyref (#15)
* Use 64bit polyref

* WIP

* Fix randomness and struct packing

* Fixes to java

* Separate some tests and ignore failing ones for now

* Remove warnings

* Fix 64bit binary issues. (#16)

* Fix 64bit binary issues.

* Fix 64bit binary issues.

* remove useless import

* Fix per AL's comment

+++ recast-java/src/main/java/io/improbable/ste/recast/FindPathResult.java
@@ -9,7 +9,7 @@ public class FindPathResult extends Structure {
     public static class ByValue extends FindPathResult implements Structure.ByValue {}
     public static class ByReference extends FindPathResult implements Structure.ByReference {}
     public int status;
-    public int path[] = new int[Const.MAX_PATH_LEN];
+    public long path[] = new long[Const.MAX_PATH_LEN];
     public int pathCount;
 
     @Override
